+++ CHANGELOG.md
@@ -43,6 +43,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#40](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/40) Timestamped public randomness commitments.
   Breaks `PubRandCommit` struct that is being used for both storage and queries.
 
+### API breaking
+
+* [#60](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/60) feat: add signing context
+
 ### State breaking
 
 * [#55](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/55) evidence: remove evidence DB